fixed branching on usize

+++ libcrux-ml-kem/src/hash_functions.rs
@@ -227,7 +227,7 @@ pub(crate) mod avx2 {
             debug_assert!(K == 2 || K == 3 || K == 4);
             let mut state = x4::incremental::shake128_init();
 
-            match K {
+            match K as u8 {
                 2 => {
                     x4::incremental::shake128_absorb_final(
                         &mut state, &input[0], &input[1], &input[0], &input[0],
@@ -255,7 +255,7 @@ pub(crate) mod avx2 {
         fn shake128_squeeze_three_blocks(&mut self) -> [[u8; THREE_BLOCKS]; K] {
             debug_assert!(K == 2 || K == 3 || K == 4);
             let mut out = [[0u8; THREE_BLOCKS]; K];
-            match K {
+            match K as u8 {
                 2 => {
                     let mut dummy_out0 = [0u8; THREE_BLOCKS];
                     let mut dummy_out1 = [0u8; THREE_BLOCKS];
@@ -301,7 +301,7 @@ pub(crate) mod avx2 {
         fn shake128_squeeze_block(&mut self) -> [[u8; BLOCK_SIZE]; K] {
             debug_assert!(K == 2 || K == 3 || K == 4);
             let mut out = [[0u8; BLOCK_SIZE]; K];
-            match K {
+            match K as u8 {
                 2 => {
                     let mut dummy_out0 = [0u8; BLOCK_SIZE];
                     let mut dummy_out1 = [0u8; BLOCK_SIZE];
@@ -385,7 +385,7 @@ pub(crate) mod neon {
         fn PRFxN<const LEN: usize>(input: &[[u8; 33]; K]) -> [[u8; LEN]; K] {
             debug_assert!(K == 2 || K == 3 || K == 4);
             let mut out = [[0u8; LEN]; K];
-            match K {
+            match K as u8 {
                 2 => {
                     let (out0, out1) = out.split_at_mut(1);
                     x2::shake256(&input[0], &input[1], &mut out0[0], &mut out1[0]);
@@ -416,7 +416,7 @@ pub(crate) mod neon {
                 x2::incremental::shake128_init(),
                 x2::incremental::shake128_init(),
             ];
-            match K {
+            match K as u8 {
                 2 => {
                     x2::incremental::shake128_absorb_final(&mut state[0], &input[0], &input[1]);
                 }
@@ -441,7 +441,7 @@ pub(crate) mod neon {
             debug_assert!(K == 2 || K == 3 || K == 4);
 
             let mut out = [[0u8; THREE_BLOCKS]; K];
-            match K {
+            match K as u8 {
                 2 => {
                     let (out0, out1) = out.split_at_mut(1);
                     x2::incremental::shake128_squeeze_first_three_blocks(
@@ -490,7 +490,7 @@ pub(crate) mod neon {
             debug_assert!(K == 2 || K == 3 || K == 4);
 
             let mut out = [[0u8; BLOCK_SIZE]; K];
-            match K {
+            match K as u8 {
                 2 => {
                     let mut out0 = [0u8; BLOCK_SIZE];
                     let mut out1 = [0u8; BLOCK_SIZE];
